+++ src/v5/shared.rs
@@ -354,7 +354,7 @@ impl Encoder for MqttShared {
 }
 
 impl Decoder for MqttShared {
-    type Item = codec::Packet;
+    type Item = (codec::Packet, u32);
     type Error = error::DecodeError;
 
     #[inline]